requirements.txt
Flask
Flask-CQLAlchemy
flask_cassandra
